refactor(memory): consolidation の責務境界を分離する
